collector/requirements.txt
websocket-client>=1.6.0
kafka-python>=2.0.2
python-dotenv>=1.0.0
fastapi==0.104.1
uvicorn==0.24.0
